fixing graphics pipeline (easy pipe)

+++ GraphicsEngine/GraphicsEngine/Source Files/Geometry/Transforms.h
@@ -3,8 +3,8 @@
 #include "../Math Utilities/vector2d.h"
 #include "Triangle.h"
 
-Vector2d Ortho2D(Vector3d vec, int zFar, int zNear, float thetaFoV);
+Vector2d Ortho2D(Vector3d vec, float zFar, float zNear, float thetaFoV);
 
-Triangle Ortho2DTris(Triangle triangle, int zFar, int zNear, float thetaFoV);
+Triangle Ortho2DTris(Triangle triangle, float zFar, float zNear, float thetaFoV);
